--- src/features/ui/skeletons/button.tsx
@@ -1,5 +1,0 @@
-import { Skeleton } from "@/shared/ui";
-
-export function ButtonSkeleton() {
-  return <Skeleton className="h-10 w-20 rounded-md" />;
-}
