+++ src/features/ui/skeletons/index.ts
@@ -1,7 +1,6 @@
+export * from "./alert";
 export * from "./carousel";
 export * from "./form";
 export * from "./tab";
 export * from "./table";
 export * from "./twidget-item";
-export * from "./button";
-export * from "./alert";
