+++ src/features/ui/skeletons/twidget-item.tsx
@@ -2,9 +2,9 @@ import { Skeleton } from "@/shared/ui";
 
 export function TwidgetItem() {
   return (
-    <div className="flex flex-col space-y-2">
+    <section className="flex flex-col space-y-2">
       <Skeleton className="h-4 w-16" />
       <Skeleton className="h-10 w-full" />
-    </div>
+    </section>
   );
 }
